ENH: Allowing to use any version of VTK that is greater or equal to 6
When finding the VTK package with cmake, we need to add the option
"NO_MODULE" to look for VTK only using the Config option[1][2]
The required version is not checked using the REQUIRED keyword anymore as
it is too restrictive (MAJOR and MINOR version of provided and requested
versions have to match). Instead, we allow any newer version than the one
that was requested, but warn the user if the version does not match
exactly.

[1] http://www.vtk.org/Bug/view.php?id=14826#c32945
[2] http://cmake.org/cmake/help/v2.8.12/cmake.html#command:find_package